+++ src/Airslip.Analytics.Core/Models/SnapshotMetric.cs
@@ -1,3 +1,3 @@
 namespace Airslip.Analytics.Core.Models;
 
-public record SnapshotMetric(long TimeStamp, long Balance);
+public record SnapshotMetric(long TimeStamp, double Balance);
